@@ -222,7 +222,7 @@ namespace EasyAbp.FileManagement.Blazor.Server.Host
 
             app.UseHttpsRedirection();
             app.UseCorrelationId();
-            app.UseStaticFiles();
+            app.MapAbpStaticAssets();
             app.UseRouting();
             app.UseAuthentication();
             app.UseJwtTokenMiddleware();
